Drag items produces undo edit, #42
* Change ArcPoint to implement PlaceabelPetriNetComponent
* Change Connectable x, y to be ints
* DragManager creates undo component by adding start and finish methods
  these save any PlaceablePetriNetComponents coordinates and copy them
  afterwards to create a MovePetriNetObject class
*

+++ pipe-core/src/main/java/pipe/io/adapters/utils/ConnectableUtils.java
@@ -61,7 +61,7 @@ public final class ConnectableUtils {
      * @param adaptedConnectable
      */
     public static void setConnectablePosition(Connectable connectable, AdaptedConnectable adaptedConnectable) {
-        connectable.setX(adaptedConnectable.getGraphics().point.getX());
-        connectable.setY(adaptedConnectable.getGraphics().point.getY());
+        connectable.setX((int)adaptedConnectable.getGraphics().point.getX());
+        connectable.setY((int)adaptedConnectable.getGraphics().point.getY());
     }
 }
